requirements.txt
psycopg2==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.1
python3-dtls==1.3.0
termcolor==2.5.0
